Embed serves different types of content.
Added a served_as parameter (optional) to embeds in order to allow them to serve
different kind of content. Current values include static (services a static map linked
through to the view page), default (the default embed) and earth (which serves a
google earth embed).

The earth embed could be linked in the future from the view page, and provides a
full kml import and display scenario.

[completed:318]